requirements.txt
# Core LangChain packages
langchain-core==0.3.37
langchain-community==0.3.18
langchain-ollama==0.2.3
langchain-openai==0.3.6

# Vector store and embeddings
chromadb==0.5.3
rank-bm25==0.2.2

# RAG evaluation
ragas==0.1.9

# Data processing and visualization
pandas==2.2.3
datasets==3.3.2
matplotlib==3.10.0
seaborn==0.13.2
tqdm==4.66.4

# API and environment
python-dotenv==1.0.1
openai==1.63.2
ollama==0.4.7
